ComplianceCheckAgent/src/ComplianceCheckAgent/agent.py

from autogen import AssistantAgent
from agoric_sdk import ContractAPI

class ComplianceCheckAgent:
    def __init__(self):
        self.assistant = AssistantAgent("compliance_assistant")
        self.contract_api = ContractAPI("https://api.agoric.net")

    def check_compliance(self, transaction):
        try:
            # Check compliance of the transaction
            compliance_result = self.contract_api.check_transaction_compliance(transaction)
            if compliance_result["status"] == "compliant":
                return "Transaction is compliant."
            else:
                return f"Transaction is not compliant: {compliance_result['reason']}"
        except Exception as e:
            return f"Failed to check compliance: {str(e)}"

    def report_non_compliance(self, transaction):
        try:
            # Report non-compliance issues
            report = self.assistant.handle_input(f"Report non-compliance for the following transaction: {transaction}")
            return report
        except Exception as e:
            return f"Failed to report non-compliance: {str(e)}"

    def generate_compliance_report(self, transactions):
        try:
            # Generate a compliance report for a list of transactions
            report = self.assistant.handle_input(f"Generate a compliance report for the following transactions: {transactions}")
            return report
        except Exception as e:
            return f"Failed to generate compliance report: {str(e)}"

if __name__ == "__main__":
    agent = ComplianceCheckAgent()
    transaction = "example_transaction"
    compliance_result = agent.check_compliance(transaction)
    print(compliance_result)
    report_result = agent.report_non_compliance(transaction)
    print(report_result)
    transactions = ["transaction1", "transaction2"]
    compliance_report = agent.generate_compliance_report(transactions)
    print(compliance_report)
